pulling out code for currentLocation into component; fix to weather key
